+++ src/layered_brownian_bridge.cpp
@@ -254,7 +254,7 @@ Rcpp::NumericMatrix layered_brownian_bridge(const double &x,
 //'
 //' @export
 // [[Rcpp::export]]
-Rcpp::NumericMatrix multi_layered_brownian_bridge(const double &dim,
+Rcpp::NumericMatrix multi_layered_brownian_bridge(const int &dim,
                                                   const Rcpp::NumericVector &x,
                                                   const Rcpp::NumericVector &y,
                                                   const double &s,
